Add Soda Cloud metrics store support (#531)
This commit enables the usage of historic metrics from Soda Cloud in your tests.
To use this feature you can specify the metrics that needs to be fetched in your
yaml for a column:

```yaml
columns:
  amount:
    metrics:
      - min
    historic_metrics:
      - name: min_avg
        type: avg
        metric: min
        count: 3
      - name: prev_min
        type: prev
        metric: min
        count: 1
    tests:
      - min == 0
      - min_avg + 1 == 1
      - prev_min == min_avg
```
In the above example, the historical metrics minimum of average and previous
minimum are fetched from Soda Cloud. The name property is used in the tests, the
metric: is the metric type. The type can be one of 'avg, min, max, or prev'. The
count specifies the number of items in the history that should be considered. In
the case of `prev` this would mean the number of measurements ago.

closes #528